update tests to use alt data
